app/src/main/java/com/nikolaswidad/storyappsnew/ui/main/MainViewModel.kt
package com.nikolaswidad.storyappsnew.ui.main

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.nikolaswidad.storyappsnew.data.StoryRepository
import com.nikolaswidad.storyappsnew.data.network.response.ListStoryItem

class MainViewModel(
    private val repository: StoryRepository
) : ViewModel() {

    fun getStory(token: String): LiveData<PagingData<ListStoryItem>> =
        repository.getStories(token).cachedIn(viewModelScope)
}